Fixed progress bar.
StreamDownloader

Changed:
-> MainWindow.xaml.cs

StreamDownloaderControls

Changed:
-> FlatWindow.cs
-> Themes/Generic.xaml

StreamDownloaderDownload

Changed:
-> Download/DownloadTask.cs
-> Download/FileDownload.cs
-> Download/JSON/StoredDownload.cs

+++ StreamDownloaderDownload/Download/DownloadTask.cs
@@ -64,7 +64,7 @@ namespace StreamDownloaderDownload.Download
             _fileDownload = new FileDownload(downloadUrl, $"{tempFolder}{Guid.NewGuid()}.dtemp", $"{downloadFolder}{fileName}.{fileType}", chunkSize, this);
         }
 
-        public DownloadTask(string downloadFolder, string tempFolder, string fileName, string fileType, string downloadUrl, uint chunkSize, uint writtenChunks, ulong contentLength)
+        public DownloadTask(string downloadFolder, string tempFolder, string fileName, string fileType, string downloadUrl, uint chunkSize, ulong writtenChunks, ulong contentLength)
         {
 
             _downloadFolder = downloadFolder;
